chore!: シンプルなディレクトリ構成に変更
