--- plugin/clipping.vim
@@ -1,1 +1,0 @@
-lua require("nvim-treesitter-clipping").init()
